pyproject.toml
[tool.poetry]
name = "dynamic-django-models"
version = "0.1.0"
description = ""
authors = ["Mite Stojanov <stojanov1990@gmail.com>"]
readme = "README.md"
packages = [{include = "dynamic_django_models"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
django-environ = "^0.10.0"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
pytest = "^7.4.0"
django-extensions = "^3.2.3"
pytest-django = "^4.5.2"
pylint-django = "^2.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
